Restructure Hello* examples: minimal Hello, HelloIotsa baseline, auth variants (#156)
Hello becomes a true single-file example again (mainHello.cpp merged back
into Hello.ino, dead uptime-logging code removed, OTA dropped) -- it's the
toe-in-the-water entry point for people coming from Arduino, and no longer
builds under platformio at all (the toplevel multi-project build_src_filter
mechanism can't compile a .ino with real code in it; only pio ci and
standalone builds can, and neither is the point of this example).

HelloApi and HelloCpp are merged into a new HelloIotsa: the minimal
reasonable normal iotsa app, with the hello module split into its own
iotsaHelloMod.h/.cpp (declaration/implementation, like iotsa's own core
modules) rather than defined inline in the main file.

HelloPasswd, HelloRights, HelloToken and HelloUser are rebased onto that
same HelloIotsa structure, each adding exactly one access-control
mechanism on top (gating both the web form and the now-added REST API via
the same authenticator passed to each module's constructor). HelloPasswd's
inline IotsaStaticAuthMod is likewise split into its own .h/.cpp.

README's sample-programs section now explains this progression explicitly,
separate from the rest of the examples (which are independent, one
capability each).
