--- examples/HelloCpp/iotsa-build.json
@@ -1,26 +1,0 @@
-{
-  "variants": [
-    {
-      "board": "nodemcuv2",
-      "ci": {
-        "platformio": true,
-        "arduino": true
-      }
-    },
-    {
-      "board": "esp32thing",
-      "ci": {
-        "platformio": true,
-        "arduino": true
-      },
-      "partitions": "min_spiffs.csv"
-    },
-    {
-      "board": "esp32c3devkit",
-      "ci": {
-        "platformio": true,
-        "arduino": false
-      }
-    }
-  ]
-}
