--- examples/HelloCpp/platformio.ini
@@ -1,29 +1,0 @@
-; PlatformIO Project Configuration File
-; Generated by extras/python/gen_build_matrix.py -- do not hand-edit.
-;
-[platformio]
-src_dir = .
-default_envs = nodemcuv2
-
-[common]
-framework = arduino
-lib_ldf_mode = deep+
-lib_deps = https://github.com/cwi-dis/iotsa.git#develop
-monitor_speed = 115200
-
-[env:nodemcuv2]
-extends = common
-platform = espressif8266
-board = nodemcuv2
-
-[env:esp32thing]
-extends = common
-platform = espressif32
-board = esp32thing
-board_build.partitions = min_spiffs.csv
-
-[env:esp32c3devkit]
-extends = common
-platform = espressif32
-board = esp32-c3-devkitm-1
-
