+++ extras/python/gen_build_matrix.py
@@ -11,6 +11,7 @@ them into:
 See docs/module-interface-status.md and the #156 discussion for background.
 """
 import argparse
+import io
 import json
 import os
 import sys
@@ -166,10 +167,12 @@ def emit_github_matrix(entries, backend, out):
 
 def emit_standalone_ini(entries, target_dir, out):
     rel = os.path.relpath(target_dir, REPO_ROOT)
-    own = [e for e in entries if e["kind"] == "example" and f"examples/{e['name']}" == rel]
+    own = [e for e in entries if e["kind"] == "example" and f"examples/{e['name']}" == rel
+           and e["ci"].get("platformio", True)]
     if not own:
-        print(f"error: no iotsa-build.json entries found for {rel}", file=sys.stderr)
-        sys.exit(1)
+        print(f"note: {rel} has no platformio-enabled variants, no standalone "
+              f"platformio.ini to generate", file=sys.stderr)
+        return
     out.write("; PlatformIO Project Configuration File\n")
     out.write("; Generated by extras/python/gen_build_matrix.py -- do not hand-edit.\n")
     out.write(";\n")
@@ -211,6 +214,27 @@ def main():
     args = ap.parse_args()
 
     entries = load_entries()
+
+    if args.format == "standalone-ini":
+        # Buffered: a directory with no platformio-enabled variants emits nothing,
+        # and any stale output file (e.g. from before a variant dropped platformio
+        # coverage) must be removed rather than left behind as an empty/stale file.
+        if not args.dir:
+            ap.error("--format=standalone-ini requires --dir")
+        buf = io.StringIO()
+        emit_standalone_ini(entries, os.path.join(REPO_ROOT, args.dir), buf)
+        content = buf.getvalue()
+        if not content:
+            if args.output and os.path.exists(args.output):
+                os.remove(args.output)
+            return
+        if args.output:
+            with open(args.output, "w") as out:
+                out.write(content)
+        else:
+            sys.stdout.write(content)
+        return
+
     out = open(args.output, "w") if args.output else sys.stdout
     try:
         if args.format == "platformio-ini":
@@ -219,10 +243,6 @@ def main():
             if not args.backend:
                 ap.error("--format=github-matrix requires --backend")
             emit_github_matrix(entries, args.backend, out)
-        elif args.format == "standalone-ini":
-            if not args.dir:
-                ap.error("--format=standalone-ini requires --dir")
-            emit_standalone_ini(entries, os.path.join(REPO_ROOT, args.dir), out)
     finally:
         if args.output:
             out.close()
