+++ generated_envs.ini
@@ -52,44 +52,18 @@ board_build.partitions = min_spiffs.csv
 extends = esp32c3devkit
 build_src_filter = +<*> +<../examples/DateTime>
 
-[env:nodemcuv2-example-Hello]
+[env:nodemcuv2-example-HelloIotsa]
 extends = nodemcuv2
-build_src_filter = +<*> +<../examples/Hello>
+build_src_filter = +<*> +<../examples/HelloIotsa>
 
-[env:esp32thing-example-Hello]
+[env:esp32thing-example-HelloIotsa]
 extends = esp32thing
-build_src_filter = +<*> +<../examples/Hello>
+build_src_filter = +<*> +<../examples/HelloIotsa>
 board_build.partitions = min_spiffs.csv
 
-[env:esp32c3devkit-example-Hello]
+[env:esp32c3devkit-example-HelloIotsa]
 extends = esp32c3devkit
-build_src_filter = +<*> +<../examples/Hello>
-
-[env:nodemcuv2-example-HelloApi]
-extends = nodemcuv2
-build_src_filter = +<*> +<../examples/HelloApi>
-
-[env:esp32thing-example-HelloApi]
-extends = esp32thing
-build_src_filter = +<*> +<../examples/HelloApi>
-board_build.partitions = min_spiffs.csv
-
-[env:esp32c3devkit-example-HelloApi]
-extends = esp32c3devkit
-build_src_filter = +<*> +<../examples/HelloApi>
-
-[env:nodemcuv2-example-HelloCpp]
-extends = nodemcuv2
-build_src_filter = +<*> +<../examples/HelloCpp>
-
-[env:esp32thing-example-HelloCpp]
-extends = esp32thing
-build_src_filter = +<*> +<../examples/HelloCpp>
-board_build.partitions = min_spiffs.csv
-
-[env:esp32c3devkit-example-HelloCpp]
-extends = esp32c3devkit
-build_src_filter = +<*> +<../examples/HelloCpp>
+build_src_filter = +<*> +<../examples/HelloIotsa>
 
 [env:nodemcuv2-example-HelloPasswd]
 extends = nodemcuv2
